main.py take care of ignored tag in the keep count
Previously the could keep was done on all tagged image but when using
IGNORE_TAGS_REGEX, the image ignored was taken in the keep count.

That means if we have 3 images

  * image1 tagfoo
  * image2 tagbar
  * image3 tag bla

and set keep 1 with ignoretags tagfoo it would have deleted image2 and
3.

Now keep is done on other image that are ignored. I the previous example
it will delete image3 only, ignore image1 (tagfoo) and  keep image2